+++ app/ExternalFiles/Files.php
@@ -1290,18 +1290,11 @@ class Files {
 	 */
 	public function add_filter_options( array $options ): array {
 		// get all external sources.
-		$terms = get_terms(
-			array(
-				'taxonomy'   => Taxonomy::get_instance()->get_name(),
-				'hide_empty' => false,
-			)
-		);
+		$terms = Directory_Listing::get_instance()->get_external_sources();
 
 		// add them to the options list.
-		if ( is_array( $terms ) ) {
-			foreach ( $terms as $term ) {
-				$options[ (string) $term->term_id ] = $term->name;
-			}
+		foreach ( $terms as $term ) {
+			$options[ (string) $term->term_id ] = $term->name;
 		}
 
 		return $options;