+++ tests/Unit/Services/Zip.php
@@ -23,4 +23,15 @@ class Zip extends externalFilesTests {
 		$this->assertIsString( $name );
 		$this->assertNotEmpty( $name );
 	}
+
+	/**
+	 * Test if the returning variable is a string.
+	 *
+	 * @return void
+	 */
+	public function test_get_directory(): void {
+		$directory = \ExternalFilesInMediaLibrary\Services\Zip::get_instance()->get_directory();
+		$this->assertIsString( $directory );
+		$this->assertEmpty( $directory );
+	}
 }